agentic_rag_LangGraph/requirements.txt
# Core dependencies
pandas>=2.0.0
pydantic>=2.0.0

# LangChain and related
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langgraph>=0.0.20

# Embeddings and reranking
sentence-transformers>=2.2.0

# Web search
tavily-python>=0.3.0

# Document processing
unstructured>=0.10.0
python-magic-bin>=0.4.14; sys_platform == 'win32'

# Database
SQLAlchemy>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# OpenAI API
openai>=1.0.0
